// src/core/plugin/iframe-editor/interface.ts
/*!
 * Jodit Editor PRO (https://xdsoft.net/jodit/)
 * See LICENSE.md in the project root for license information.
 * Copyright (c) 2013-2022 Valeriy Chupurnov. All rights reserved. https://xdsoft.net/jodit/pro/
 */

export interface IframeEditorState {
	src: string;
	className: string;
	width: number;
	height: number;
	enableScrollbars: boolean;
	frameBorder: boolean;
	name: string;
	title: string;
}